+++ src/Enums/TipoDocumentoAsociado.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Nyxcode\PhpSifenTool\Enums;
+
+enum TipoDocumentoAsociado: int
+{
+    case ELECTRONICO = 1;
+    case IMPRESO = 2;
+    case CONSTANCIA_ELECTRONICA = 3;
+
+    public function getDescripcion(): string
+    {
+        return match ($this) {
+            self::ELECTRONICO => 'Electrónico',
+            self::IMPRESO => 'Impreso',
+            self::CONSTANCIA_ELECTRONICA => 'Constancia Electrónica',
+        };
+    }
+}