+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added tag classes for group E.7.2.1
 -   Added tag classes for group E.8
 -   Added tag classes for group E.8.1
+-   Added tag classes for group E.8.1.1
+-   Added tag classes for group E.8.2
 -   Base classes of the Builder pattern are added for the generation of Factura Electronica
 -   Factura Electronica generation test added